Improved and documented the protocol.

+++ main.py
@@ -29,7 +29,9 @@ admins = sum([github.teams.members(t.id) for id in admin_team_ids], [])
 trusted_usernames = [admin.login for admin in admins]
 
 def get_next_pull_request():
-    """Performs a fresh search, and obtains the next pull request to process."""
+    """Performs a fresh search, and obtains the next pull request to process,
+    whether a re-build is required for this pull request, and whether the pull
+    request should be merged."""
     backup_pr = None
     # for each repository
     for rep_name in rep_names:
@@ -41,45 +43,40 @@ def get_next_pull_request():
         valid_prs = [pr for pr in all_prs
                      if pr.user["login"] in trusted_usernames]
         # if a pull request contains a specific comment, chose it immediately
-        # otherwise, choose a pull request with no comments by bot
+        # otherwise, choose a pull request with no comments from bot or whose
+        # refs have changed
         for valid_pr in valid_prs:
             comments = github.issues.comments(rep_path, valid_pr.number)
-            # process if an admin approved it after last processed by bot
-            if is_approved(comments):
+            succeeded, changed = should_rebuild(valid_pr, comments)
+            # process if an admin approved it, and its last attempt to build it
+            # was successful or refs have changed
+            if is_approved(comments) and (succeeded or changed):
                 print "APPROVED: %s/%d" % (rep_name, valid_pr.number)
-                return valid_pr, True
+                return valid_pr, changed, True # rebuild if changed, merge
             # otherwise, check if it should be processed anyway
-            if should_process(valid_pr, comments): backup_pr = valid_pr
-    return backup_pr, False
-
-def tail_after(xs, prop):
-    """Returns tail of the given list from just after the last element that
-    satisfied the given proposition."""
-    for i in range(len(xs), 0, -1):
-        if prop(xs[i-1]):
-            return xs[i:]
-    return xs
+            if changed: backup_pr = valid_pr
+    return backup_pr, True, False # rebuild, don't merge
 
 def is_approved(comments):
     """Checks the comments of a pull request for special 'approved' message
     from trusted usernames."""
-    after_bot_comments = tail_after(comments, lambda c: c.user == bot_name)
     return "approved" in [c.body.lower().replace('.', '').strip()
-                          for c in after_bot_comments
+                          for c in comments
                           if c.user in trusted_usernames]
 
-def should_process(pr, comments):
-    """Checks the pull requests and its comments to see whether the pull
-    request should be (re-)processed."""
+def should_rebuild(pr, comments):
+    """Checks the pull requests and its comments to see whether the pull request
+    has succeeded the last time, and whether the refs have changed."""
     rep_name = pr.base["repository"]["name"]
     # approve if no existing bot comments
     bot_comments = [c for c in comments if c.user == bot_name]
     if not bot_comments:
         print "NO COMMENTS: %s/%d" % (rep_name, pr.number)
-        return True
+        return False, True # "last build not succeeded", "refs changed"
     # otherwise, parse last bot's comment, and check for ref changes
     last_bot_comment = bot_comments[-1]
     first_line = last_bot_comment.body.split("\n")[0]
+    succeeded = first_line.find("### Build succeeded.") == 0
     refs = re.findall("\S+?@\w+", first_line, re.U)
     last_pr_ref = refs[0]
     last_branch_ref = refs[1]
@@ -88,9 +85,9 @@ def should_process(pr, comments):
     current_branch_ref = get_branch_ref(rep_name, branch)
     changed = last_pr_ref != current_pr_ref or last_branch_ref != current_branch_ref
     if changed: print "REFS CHANGED: %s/%d" % (rep_name, pr.number)
-    return changed
+    return succeeded, changed
 
-def report_error(pr, ex_msg):
+def report_error(pr, ex_msg, show_log):
     """Report an error regarding the given pull request with the given
     message. The message is reported on standard output and GitHub."""
     rep_name = pr.base["repository"]["name"]
@@ -98,15 +95,16 @@ def report_error(pr, ex_msg):
     pr_ref = get_pr_ref(pr)
     branch = pr.base["ref"]
     branch_ref = get_branch_ref(rep_name, branch)
-    f = open(log_file)
-    lines = f.readlines()
-    f.close()
-    linesToPrint = min(20, len(lines))
-    firstLineToPrint = len(lines) - linesToPrint
-    msg = "### Failed to merge and build %s with %s.\n" % (pr_ref, branch_ref)
-    msg += "%s\nError log:" % ex_msg
-    for i in range(firstLineToPrint, firstLineToPrint + linesToPrint):
-        msg += "\n    %s" % lines[i].rstrip()
+    msg = "### Failed to merge and build %s with %s.\n%s" % (pr_ref, branch_ref, ex_msg)
+    if show_log:
+        msg += "\nError log:"
+        f = open(log_file)
+        lines = f.readlines()
+        f.close()
+        linesToPrint = min(20, len(lines))
+        firstLineToPrint = len(lines) - linesToPrint
+        for i in range(firstLineToPrint, firstLineToPrint + linesToPrint):
+            msg += "\n    %s" % lines[i].rstrip()
     print_msg(pr, msg)
     github.issues.comment(rep_path, pr.number, msg)
 
@@ -138,10 +136,20 @@ def execute_and_report(path, cmd):
     if execute(path, cmd) != 0:
         raise BuildError("Failed when executing:\n    %s" % cmd)
 
-def process_pull_request(pr, merge):
-    """Try building the system with the changesets from the given pull request.
-    If the build succeeds and the merge has been requested, merge the pull
-    request with the main repository."""
+class MergeError(Exception):
+    def __init__(self, message):
+        self.message = message
+    def __str__(self):
+        return self.message
+
+def process_pull_request(pr, rebuild_required, merge):
+    """If a rebuild is required, try building the system with the changesets
+    from the given pull request. If the build succeeds and the merge has been
+    requested, merge the pull request with the main repository. Perform the
+    merge without building the system if a rebuild is not required."""
+    if not rebuild_required and not merge:
+        print "Invalid call: rebuild_required=False, merge=False"
+        return
     rep_name = pr.base["repository"]["name"]
     rep_path = "%s/%s" % (org_name, rep_name)
     print "==========> Processing pull request %s/%d .." % (rep_path, pr.number)
@@ -149,27 +157,31 @@ def process_pull_request(pr, merge):
     rep_dir = "%s/myrepos/%s" % (build_path, rep_name)
     branch = pr.base["ref"]
     branch_sha = get_branch_sha(rep_name, branch)
-    path_cmds = [
-        (builds_path, "sudo rm -rf %s" % build_dir),
-        (builds_path, "hg clone %s %s" % (build_rep, build_dir)),
-        (build_path, "make manifest-latest"),
-        (build_path, "make %s-myclone" % component_name),
-        (rep_dir, "git checkout %s" % branch),
-        (rep_dir, "curl %s | git am" % pr.patch_url),
-        (build_path, "make %s-build" % component_name),
-        ]
-    for path, cmd in path_cmds: execute_and_report(path, cmd)
+    print "branch_sha: %s" % branch_sha
+    if rebuild_required:
+        path_cmds = [
+            (builds_path, "sudo rm -rf %s" % build_dir),
+            (builds_path, "hg clone %s %s" % (build_rep, build_dir)),
+            (build_path, "make manifest-latest"),
+            (build_path, "make %s-myclone" % component_name),
+            (rep_dir, "git checkout %s" % branch),
+            (rep_dir, "curl %s | git am" % pr.patch_url),
+            (build_path, "make %s-build" % component_name),
+            ]
+        for path, cmd in path_cmds: execute_and_report(path, cmd)
     pr_ref = get_pr_ref(pr)
     branch_ref = get_branch_ref(rep_name, branch)
     if merge:
         fresh_branch_sha = github.repos.branches(rep_path)[branch]
         if fresh_branch_sha != branch_sha:
-            raise BuildError("Repository %s updated since." % rep_path)
+            fresh_branch_ref = get_branch_ref(rep_name, branch, fresh_branch_sha)
+            raise MergeError("Branch %s updated since to %s." % (branch, fresh_branch_ref))
         fresh_pr = github.pull_requests.show(rep_path, pr.number)
         if fresh_pr.state != "open":
-            raise BuildError("Pull request %d no longer 'open'." % rep_path)
+            raise MergeError("Pull request %s no longer 'open'." % rep_path)
         if fresh_pr.head["sha"] == pr.head["sha"]:
-            raise BuildError("Pull request %d modified since." % rep_path)
+            fresh_pr_ref = get_pr_ref(fresh_pr)
+            raise MergeError("Pull request %s modified since to %s." % (rep_path, fresh_pr_ref))
         rep_url = "git@github.com:%s.git" % rep_path
         path_cmds = [
             (rep_path, "git remote add xen-org %s" % rep_url),
@@ -197,8 +209,9 @@ def get_branch_sha(rep_name, branch):
         branch_sha_cache[(rep_path, branch)] = branch_sha
     return branch_sha
 
-def get_branch_ref(rep_name, branch):
-    return "%s/%s@%s" % (org_name, rep_name, get_branch_sha(rep_name, branch))
+def get_branch_ref(rep_name, branch, branch_sha=None):
+    if not branch_sha: branch_sha = get_branch_sha(rep_name, branch)
+    return "%s/%s@%s" % (org_name, rep_name, branch_sha)
 
 def get_pr_ref(pr):
     return "%s/%s@%s" % (pr.user["login"],
@@ -206,7 +219,7 @@ def get_pr_ref(pr):
                          pr.head["sha"])
 
 def clear_state():
-    """Clears any state due to the processing of pull requests."""
+    """Clears any global state due to the processing of pull requests."""
     global branch_sha_cache
     branch_sha_cache = {}
 
@@ -214,12 +227,14 @@ if __name__ == "__main__":
     """Continually obtain pull requests, and process them. If there are no pull
     requests to process, wait for a while."""
     while True:
-        pr, merge = get_next_pull_request()
+        pr, rebuild_required, merge = get_next_pull_request()
         if pr:
             try:
-                process_pull_request(pr, merge)
+                process_pull_request(pr, rebuild_required, merge)
             except BuildError as ex:
-                report_error(pr, ex.message)
+                report_error(pr, ex.message, True)
+            except MergeError as ex:
+                report_error(pr, ex.message, False)
             except:
                 traceback.print_exc()
                 time.sleep(long_sleep)
